fix(docs): Move `NativeScriptError` declaration to `application.d.ts`
It is being used there and needs to be documented as well. The reason
that it wasn't included in the documentation by now is that `tns-core-modules.d.ts`
(which imports `module.d.ts`) defines types which are part of the internal
modules and runtimes APIs. As such they are excluded from the generation
of API documentation.